move `lauch_threads` into `process` mod
